Show form again if user enters invalid password
**Why**: That is the expected behavior. There was a bug where entering
an invalid password on the update password form would result in an
exception because the `edit` and `update` actions were using different
models.

**How**: Consolidate UpdateUserPassword and UpdateUserPasswordForm
and prefer putting it in `/forms` vs `/services` for consistency.
